ExceptionToAssertionPass moves to SSCP and first cleanups
