+++ src/compiler/AdaptiveCppLlvmPasses.cpp
@@ -37,10 +37,6 @@
 #include "hipSYCL/compiler/reflection/FunctionNameExtractionPass.hpp"
 #endif
 
-#ifdef HIPSYCL_WITH_DEVICE_EXCEPTIONS
-#include "hipSYCL/compiler/reflection/ExceptionToAssertionPass.hpp"
-// #include "hipSYCL/compiler/reflection/FunctionNameExtractionPass.hpp"
-#endif
 
 #include "llvm/Pass.h"
 #include "llvm/Passes/PassBuilder.h"
@@ -172,15 +168,6 @@ return {
                   MPM.addPass(FunctionNameExtractionPass{});
                 });
 #endif
-
-/*
-#ifdef HIPSYCL_WITH_DEVICE_EXCEPTIONS
-          PB.registerPipelineStartEPCallback(
-            [&](llvm::ModulePassManager &MPM, OptLevel Level) {
-              MPM.addPass(ExceptionToAssertionPass{});
-            });
-#endif
-*/
 
 #ifdef HIPSYCL_WITH_STDPAR_COMPILER
           if(EnableStdPar) {
